requirements.txt
wheel
termcolor
schedule
kittentts @ https://github.com/KittenML/KittenTTS/releases/download/0.8.1/kittentts-0.8.1-py3-none-any.whl
soundfile
prettytable
webdriver_manager
selenium_firefox
selenium
ollama
moviepy
Pillow>=10.0.0
yagmail
assemblyai
faster-whisper
srt_equalizer
undetected_chromedriver
platformdirs
google-generativeai
